+++ core/src/org/zapylaev/sandbox/renderer/LineRenderer.java
@@ -4,7 +4,6 @@ import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import org.zapylaev.sandbox.map.DiamondSquareGenerator;
-import org.zapylaev.sandbox.map.GenerateFromFile;
 import org.zapylaev.sandbox.map.MapGenerator;
 
 /**
@@ -13,7 +12,7 @@ import org.zapylaev.sandbox.map.MapGenerator;
 public class LineRenderer implements Renderer {
 
     private final MapGenerator mMapGenerator;
-    private int[] mLine;
+    private double[] mLine;
     private ShapeRenderer mShapeRenderer;
 
     public LineRenderer() {
@@ -28,11 +27,11 @@ public class LineRenderer implements Renderer {
         mShapeRenderer.begin(ShapeRenderer.ShapeType.Line);
         mShapeRenderer.setColor(Color.RED);
         for (int i = 0; i < mLine.length; i++) {
-            float y = mLine[i] * 10;
+            float y = (float) (mLine[i] * 10);
             float x = i * 10;
             mShapeRenderer.line(x, 0, x, y);
             if (i < mLine.length - 1) {
-                mShapeRenderer.line(x, y, (i + 1) * 10, mLine[i + 1] * 10);
+                mShapeRenderer.line(x, y, (i + 1) * 10, (float) (mLine[i + 1]) * 10);
             }
         }
         mShapeRenderer.end();
